:(
With more testing we realised that you have to clear
table in between each search or results get crapped
(uses static evaluation too early in search).

After this we pretty much lost all speed benefit...
Not sure if this is worth having since I'm not 100%
convinced it doesn't affect the search.