Python script is GO
Yeah, it's ugly ugly and hacky as shit get off my back.

Anyway I think the charts look rad.
I BELIEVE that ziggy zag ness is due to parity; different weights
are good depending on whether or not you will be getting the last move.
Or it's a bug. Either way; the former will sound good in the report.